+++ src/rust/src/declarative_asn1/types.rs
@@ -254,6 +254,10 @@ impl PrintableString {
         (**self.inner.bind(py)).eq(other.inner.bind(py))
     }
 
+    fn __hash__(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<isize> {
+        (**self.inner.bind(py)).hash()
+    }
+
     pub fn __repr__<'py>(
         &self,
         py: pyo3::Python<'py>,
@@ -299,6 +303,10 @@ impl IA5String {
 
     fn __eq__(&self, py: pyo3::Python<'_>, other: pyo3::PyRef<'_, Self>) -> pyo3::PyResult<bool> {
         (**self.inner.bind(py)).eq(other.inner.bind(py))
+    }
+
+    fn __hash__(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<isize> {
+        (**self.inner.bind(py)).hash()
     }
 
     pub fn __repr__<'py>(
@@ -359,6 +367,10 @@ impl UtcTime {
         (**self.inner.bind(py)).eq(other.inner.bind(py))
     }
 
+    fn __hash__(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<isize> {
+        (**self.inner.bind(py)).hash()
+    }
+
     pub fn __repr__<'py>(
         &self,
         py: pyo3::Python<'py>,
@@ -402,6 +414,10 @@ impl GeneralizedTime {
 
     fn __eq__(&self, py: pyo3::Python<'_>, other: pyo3::PyRef<'_, Self>) -> pyo3::PyResult<bool> {
         (**self.inner.bind(py)).eq(other.inner.bind(py))
+    }
+
+    fn __hash__(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<isize> {
+        (**self.inner.bind(py)).hash()
     }
 
     pub fn __repr__<'py>(
@@ -451,6 +467,12 @@ impl BitString {
     fn __eq__(&self, py: pyo3::Python<'_>, other: pyo3::PyRef<'_, Self>) -> pyo3::PyResult<bool> {
         Ok((**self.data.bind(py)).eq(other.data.bind(py))?
             && self.padding_bits == other.padding_bits)
+    }
+
+    fn __hash__(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<isize> {
+        (self.data.bind(py), self.padding_bits)
+            .into_pyobject(py)?
+            .hash()
     }
 
     pub fn __repr__<'py>(